Move all data to new namespace
