+++ .Lib9c.Tests/Action/AdventureBoss/ClaimAdventureBossRewardTest.cs
@@ -10,7 +10,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
     using Nekoyume;
     using Nekoyume.Action;
     using Nekoyume.Action.AdventureBoss;
-    using Nekoyume.Helper;
+    using Nekoyume.Data;
     using Nekoyume.Model.State;
     using Nekoyume.Module;
     using Nekoyume.TableData;
@@ -109,7 +109,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             yield return new object[]
             {
                 0, 100, false, null,
-                new ClaimableReward
+                new AdventureBossData.ClaimableReward
                 {
                     NcgReward = 5 * NCG, // 5% of 100 NCG
                     ItemReward = new Dictionary<int, int>
@@ -129,7 +129,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             yield return new object[]
             {
                 1, 100, true, 100,
-                new ClaimableReward
+                new AdventureBossData.ClaimableReward
                 {
                     NcgReward = 10 * NCG, // 5% of 200 NCG
                     ItemReward = new Dictionary<int, int>
@@ -149,7 +149,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             yield return new object[]
             {
                 3, 100, true, 200,
-                new ClaimableReward
+                new AdventureBossData.ClaimableReward
                 {
                     NcgReward = 15 * NCG, // 5% of 300 NCG
                     ItemReward = new Dictionary<int, int>
@@ -171,7 +171,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
         {
             yield return new object[]
             {
-                0, 100, 0, new ClaimableReward
+                0, 100, 0, new AdventureBossData.ClaimableReward
                 {
                     NcgReward = (5 + 15) * NCG, // 5NCG for raffle, 15NCG for 15% distribution
                     ItemReward = new Dictionary<int, int>
@@ -190,7 +190,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
 
             yield return new object[]
             {
-                0, 100, 1, new ClaimableReward
+                0, 100, 1, new AdventureBossData.ClaimableReward
                 {
                     NcgReward =
                         // 5NCG for raffle, 7.5NCG for half of 15% distribution
@@ -214,7 +214,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
 
             yield return new object[]
             {
-                1, 100, 1, new ClaimableReward
+                1, 100, 1, new AdventureBossData.ClaimableReward
                 {
                     // No raffle, 7.5 NCG for half of 15% distribution
                     NcgReward = FungibleAssetValue.FromRawValue(NCG, (BigInteger)(7.5 * 100)),
@@ -236,7 +236,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
 
             yield return new object[]
             {
-                0, 200, 1, new ClaimableReward
+                0, 200, 1, new AdventureBossData.ClaimableReward
                 {
                     // 10NCG for raffle, 15NCG for half of 15% distribution
                     NcgReward = (10 + 15) * NCG,
@@ -256,7 +256,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
 
             yield return new object[]
             {
-                1, 100, 2, new ClaimableReward
+                1, 100, 2, new AdventureBossData.ClaimableReward
                 {
                     // No raffle, 5 NCG for 1/3 of 15% distribution
                     NcgReward = 5 * NCG,
@@ -279,7 +279,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
         {
             yield return new object[]
             {
-                true, false, new ClaimableReward
+                true, false, new AdventureBossData.ClaimableReward
                 {
                     NcgReward = 5 * NCG, // 5NCG for raffle
                     ItemReward = new Dictionary<int, int>
@@ -297,7 +297,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             };
             yield return new object[]
             {
-                false, true, new ClaimableReward
+                false, true, new AdventureBossData.ClaimableReward
                 {
                     NcgReward = 20 * NCG, // 5NCG for raffle, 15NCG for 15% distribution
                     ItemReward = new Dictionary<int, int>
@@ -315,7 +315,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             };
             yield return new object[]
             {
-                true, true, new ClaimableReward
+                true, true, new AdventureBossData.ClaimableReward
                 {
                     // 5NCG for wanted raffle, 5NCG for explore raffle, 15NCG for 15% distribution
                     NcgReward = 25 * NCG,
@@ -334,7 +334,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             };
             yield return new object[]
             {
-                false, false, new ClaimableReward
+                false, false, new AdventureBossData.ClaimableReward
                 {
                     NcgReward = 0 * NCG,
                     ItemReward = new Dictionary<int, int>
@@ -359,7 +359,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             int bounty,
             bool anotherWanted,
             int anotherBounty,
-            ClaimableReward expectedReward
+            AdventureBossData.ClaimableReward expectedReward
         )
         {
             // Settings
@@ -432,7 +432,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
         {
             const int seed = 0;
             // Settings
-            var expectedReward = new ClaimableReward
+            var expectedReward = new AdventureBossData.ClaimableReward
             {
                 NcgReward = 10 * NCG,
                 FavReward = new Dictionary<int, int>
@@ -533,7 +533,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             int seed,
             int bounty,
             int anotherExplorerCount,
-            ClaimableReward expectedReward
+            AdventureBossData.ClaimableReward expectedReward
         )
         {
             // Settings
@@ -638,7 +638,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
         {
             const int seed = 0;
             // Settings
-            var expectedReward = new ClaimableReward
+            var expectedReward = new AdventureBossData.ClaimableReward
             {
                 // (5NCG for raffle, 15NCG for 15% distribution) for season 1 and 3
                 NcgReward = 40 * NCG,
@@ -774,7 +774,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
         {
             const int seed = 0;
             // Settings
-            var expectedReward = new ClaimableReward
+            var expectedReward = new AdventureBossData.ClaimableReward
             {
                 NcgReward =
                     25 * NCG, // 5NCG for wanted raffle, 5NCG for explore raffle, 15NCG for 15% distribution.
@@ -855,7 +855,11 @@ namespace Lib9c.Tests.Action.AdventureBoss
 
         [Theory]
         [MemberData(nameof(GetPrevRewardTestData))]
-        public void PrevReward(bool wanted, bool explore, ClaimableReward expectedReward)
+        public void PrevReward(
+            bool wanted,
+            bool explore,
+            AdventureBossData.ClaimableReward expectedReward
+        )
         {
             // Settings
             const int seed = 0;
@@ -955,7 +959,7 @@ namespace Lib9c.Tests.Action.AdventureBoss
             return state;
         }
 
-        private void Test(IWorld world, ClaimableReward expectedReward)
+        private void Test(IWorld world, AdventureBossData.ClaimableReward expectedReward)
         {
             Assert.Equal(expectedReward.NcgReward, world.GetBalance(TesterAddress, NCG));
             foreach (var fav in expectedReward.FavReward)
