+++ Lib9c/Helper/AdventureBossHelper.cs
@@ -9,17 +9,12 @@ using Libplanet.Crypto;
 using Libplanet.Types.Assets;
 using Nekoyume.Action;
 using Nekoyume.Action.AdventureBoss;
+using Nekoyume.Data;
 using Nekoyume.Model.AdventureBoss;
 using Nekoyume.Module;
 
 namespace Nekoyume.Helper
 {
-    public struct ClaimableReward
-    {
-        public FungibleAssetValue? NcgReward;
-        public Dictionary<int, int> ItemReward;
-        public Dictionary<int, int> FavReward;
-    }
 
 
     public static class AdventureBossHelper
@@ -34,21 +29,6 @@ namespace Nekoyume.Helper
         public const decimal TotalRewardMultiplier = 1.2m;
         public const decimal FixedRewardRatio = 0.7m;
         public const decimal RandomRewardRatio = 1m - FixedRewardRatio;
-        public const decimal NcgRuneRatio = 2.5m;
-
-        public static readonly ImmutableDictionary<int, decimal> NcgRewardRatio =
-            new Dictionary<int, decimal>
-            {
-                { 600201, 0.5m }, // Golden Dust : 0.5
-                { 600202, 1.5m }, // Ruby Dust : 1.5
-                { 600203, 7.5m }, // Diamond Dust : 7.5
-                { 600301, 0.1m }, // Normal Hammer : 0.1
-                { 600302, 0.5m }, // Rare Hammer : 0.5
-                { 600303, 1m }, // Epic Hammer : 1
-                { 600304, 4m }, // Unique Hammer : 4
-                { 600305, 15m }, // Legendary Hammer : 15
-                // { 600306, 60m }, // Divinity Hammer : 60
-            }.ToImmutableDictionary();
 
         public static (int?, int?) PickReward(IRandom random, Dictionary<int, int> itemIdDict,
             Dictionary<int, int> favIdDict)
@@ -86,7 +66,8 @@ namespace Nekoyume.Helper
             return (itemId, favId);
         }
 
-        private static ClaimableReward AddReward(ClaimableReward reward, bool isItem, int id,
+        private static AdventureBossData.ClaimableReward AddReward(
+            AdventureBossData.ClaimableReward reward, bool isItem, int id,
             int amount)
         {
             if (amount == 0)
@@ -175,8 +156,9 @@ namespace Nekoyume.Helper
             return states;
         }
 
-        public static ClaimableReward CalculateWantedReward(
-            ClaimableReward reward, BountyBoard bountyBoard, Address avatarAddress,
+        public static AdventureBossData.ClaimableReward CalculateWantedReward(
+            AdventureBossData.ClaimableReward reward, BountyBoard bountyBoard,
+            Address avatarAddress,
             out FungibleAssetValue ncgReward
         )
         {
@@ -206,15 +188,17 @@ namespace Nekoyume.Helper
             var totalFixedRewardNcg = (int)Math.Round(totalRewardNcg * FixedRewardRatio);
             var totalFixedRewardAmount = (int)Math.Round(
                 bountyBoard.FixedRewardItemId is not null
-                    ? totalFixedRewardNcg / NcgRewardRatio[(int)bountyBoard.FixedRewardItemId]
-                    : totalFixedRewardNcg / NcgRuneRatio
+                    ? totalFixedRewardNcg /
+                      AdventureBossData.NcgRewardRatio[(int)bountyBoard.FixedRewardItemId]
+                    : totalFixedRewardNcg / AdventureBossData.NcgRuneRatio
             );
 
             var totalRandomRewardNcg = (int)Math.Round(totalRewardNcg * RandomRewardRatio);
             var totalRandomRewardAmount = (int)Math.Round(
                 bountyBoard.RandomRewardItemId is not null
-                    ? totalRandomRewardNcg / NcgRewardRatio[(int)bountyBoard.RandomRewardItemId]
-                    : totalRandomRewardNcg / NcgRuneRatio
+                    ? totalRandomRewardNcg /
+                      AdventureBossData.NcgRewardRatio[(int)bountyBoard.RandomRewardItemId]
+                    : totalRandomRewardNcg / AdventureBossData.NcgRuneRatio
             );
 
             // Calculate my reward
@@ -244,10 +228,11 @@ namespace Nekoyume.Helper
         }
 
         public static IWorld CollectWantedReward(IWorld states, IActionContext context,
-            ClaimableReward reward, long currentBlockIndex, long season, Address avatarAddress,
-            out ClaimableReward collectedReward)
+            AdventureBossData.ClaimableReward reward, long currentBlockIndex, long season,
+            Address avatarAddress,
+            out AdventureBossData.ClaimableReward collectedReward)
         {
-            var agentAddress= states.GetAvatarState(avatarAddress).agentAddress;
+            var agentAddress = states.GetAvatarState(avatarAddress).agentAddress;
             for (var szn = season; szn > 0; szn--)
             {
                 var seasonInfo = states.GetSeasonInfo(szn);
@@ -299,7 +284,8 @@ namespace Nekoyume.Helper
             return states;
         }
 
-        public static ClaimableReward CalculateExploreReward(ClaimableReward reward,
+        public static AdventureBossData.ClaimableReward CalculateExploreReward(
+            AdventureBossData.ClaimableReward reward,
             BountyBoard bountyBoard, ExploreBoard exploreBoard,
             Explorer explorer, Address avatarAddress, out FungibleAssetValue ncgReward)
         {
@@ -340,8 +326,8 @@ namespace Nekoyume.Helper
 
             // calculate total reward
             var ncgRewardRatio = exploreBoard.FixedRewardItemId is not null
-                ? NcgRewardRatio[(int)exploreBoard.FixedRewardItemId]
-                : NcgRuneRatio;
+                ? AdventureBossData.NcgRewardRatio[(int)exploreBoard.FixedRewardItemId]
+                : AdventureBossData.NcgRuneRatio;
             var totalRewardAmount = (int)Math.Round(exploreBoard.UsedApPotion / ncgRewardRatio);
             reward = AddReward(reward, exploreBoard.FixedRewardItemId is not null,
                 (int)(exploreBoard.FixedRewardItemId ?? exploreBoard.FixedRewardFavId)!,
@@ -352,9 +338,10 @@ namespace Nekoyume.Helper
             return reward;
         }
 
-        public static IWorld CollectExploreReward( IWorld states, IActionContext context,
-            ClaimableReward reward, long currentBlockIndex, long season, Address avatarAddress,
-            out ClaimableReward collectedReward)
+        public static IWorld CollectExploreReward(IWorld states, IActionContext context,
+            AdventureBossData.ClaimableReward reward, long currentBlockIndex, long season,
+            Address avatarAddress,
+            out AdventureBossData.ClaimableReward collectedReward)
         {
             var agentAddress = states.GetAvatarState(avatarAddress).agentAddress;
             for (var szn = season; szn > 0; szn--)
